Update Arrayalpha
Q5. Write a C# program that declares an array alpha of 50 components of the type float Initialize the array so that the first 25 components are equal to the square of the index variable and the last 25 components are equal to three times the index variable.Output the array so that 10 elements per line are printed.